requirements.txt
numpy==1.17.2
matplotlib==3.1.1
opencv-python==4.1.1.26
pillow==6.1.0
scipy==1.3.1